Added deserializer class responsible for checking encoded uri version
